+++ src/ScottPlot5/ScottPlot5/Plottables/FunctionPlot.cs
@@ -1,27 +1,17 @@
 ﻿namespace ScottPlot.Plottables;
 
-public class FunctionPlot : IPlottable
+public class FunctionPlot(IFunctionSource source) : IPlottable
 {
     public bool IsVisible { get; set; } = true;
     public IAxes Axes { get; set; } = new Axes();
-
-    /// <summary>
-    /// Default horizontal range to use when autoscaling
-    /// </summary>
-    public CoordinateRange XRange { get; set; } = new(-10, 10);
-
     public string? Label { get; set; }
     public LineStyle LineStyle { get; } = new();
-    IFunctionSource Source { get; set; }
-    AxisLimits LastRenderLimits { get; set; } = AxisLimits.NoLimits;
+    private IFunctionSource Source { get; set; } = source;
+    private CoordinateRange MaxObservedRangeY { get; set; } = CoordinateRange.NotSet;
+    private CoordinateRange LastRenderHorizontalSpan { get; set; } = new(-10, 10);
 
-    public FunctionPlot(IFunctionSource source)
-    {
-        Source = source;
-    }
     private double MinX => Math.Max(Source.RangeX.Min.FiniteCoallesce(Axes.XAxis.Min), Axes.XAxis.Min);
     private double MaxX => Math.Min(Source.RangeX.Max.FiniteCoallesce(Axes.XAxis.Max), Axes.XAxis.Max);
-
 
     public IEnumerable<LegendItem> LegendItems => EnumerableExtensions.One(
         new LegendItem
@@ -33,7 +23,12 @@ public class FunctionPlot : IPlottable
 
     public AxisLimits GetAxisLimits()
     {
-        return LastRenderLimits;
+        if (MaxObservedRangeY == CoordinateRange.NotSet)
+            return new AxisLimits(-10, 10, -10, 10);
+
+        bool sourceRangeIsFinite = !(double.IsInfinity(Source.RangeX.Min) || double.IsInfinity(Source.RangeX.Max));
+        CoordinateRange xRange = sourceRangeIsFinite ? Source.RangeX : LastRenderHorizontalSpan;
+        return new AxisLimits(xRange, MaxObservedRangeY);
     }
 
     public void Render(RenderPack rp)
@@ -75,6 +70,16 @@ public class FunctionPlot : IPlottable
 
         rp.Canvas.DrawPath(path, paint);
 
-        LastRenderLimits = new AxisLimits(XRange.Min, XRange.Max, min, max);
+        bool somethingWasRendered = min <= max;
+        if (somethingWasRendered)
+        {
+            if (min < MaxObservedRangeY.Min)
+                MaxObservedRangeY = MaxObservedRangeY.Expanded(min);
+
+            if (max > MaxObservedRangeY.Max)
+                MaxObservedRangeY = MaxObservedRangeY.Expanded(max);
+        }
+
+        LastRenderHorizontalSpan = Axes.XAxis.Range.ToCoordinateRange;
     }
 }